fix(Dataset enums): refactor and move them to the same file
