--- src/dataset/domain/models/LabelSemanticMeaning.enum.ts
@@ -1,8 +1,0 @@
-export enum LabelSemanticMeaning {
-  DATASET = 'dataset',
-  FILE = 'file',
-  SUCCESS = 'success',
-  INFO = 'info',
-  WARNING = 'warning',
-  DANGER = 'danger'
-}
