--- src/dataset/domain/models/MetadataBlockName.ts
@@ -1,18 +1,0 @@
-export enum MetadataBlockName {
-  CITATION = 'citation',
-  GEOSPATIAL = 'geospatial',
-  ASTROPHYSICS = 'astrophysics',
-  BIOMEDICAL = 'biomedical',
-  CODE_META = 'codeMeta20',
-  COMPUTATIONAL_WORKFLOW = 'computationalWorkflow',
-  CUSTOM_ARCS = 'customARCS',
-  CUSTOM_CHIA = 'customCHIA',
-  CUSTOM_DIGAAI = 'customDigaai',
-  CUSTOM_GSD = 'customGSD',
-  CUSTOM_MRA = 'customMRA',
-  CUSTOM_PSI = 'customPSI',
-  CUSTOM_PSRI = 'customPSRI',
-  CUSTOM_HBGDKI = 'custom_hbgdki',
-  JOURNALS = 'journals',
-  SOCIAL_SCIENCE = 'socialscience'
-}
